refactor(solver): extract smt_types.py for type system
Extract type system components from the 1860-line smt.py into a focused
smt_types.py module. This includes Option types, sort registry, type
inference, temporal helpers, and encode_literal. Pure refactoring with
no behavior change — all 46 tests pass.
